% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\docType{data}
\name{locations}
\alias{locations}
\title{Sensor locations at Queen Elizabeth Central Hospital
(QECH) in Blantyre, Malawi.}
\format{
A tibble with 9 rows and 4 variables:
\describe{
\item{location}{Descriptive name of the location.}
\item{long}{Longitude coordinate.}
\item{lat}{Latitude coordinate.}
\item{type}{Type of location.}
}
}
\usage{
locations
}
\description{
Sensor locations at Queen Elizabeth Central Hospital
(QECH) in Blantyre, Malawi.
}
\keyword{datasets}
