% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qechairquality-package.R
\docType{package}
\name{qechairquality-package}
\alias{qechairquality}
\alias{qechairquality-package}
\title{qechairquality: Air Quality Data at Queen Elizabeth Central Hospital (QECH), Blantyre, Malawi}
\description{
Air quality data with measurements in 5-minute intervals for particulate matter (PM 2.5 & PM 10) at eight sensor locations over two months at Queen Elizabeth Central Hospital (QECH) in Blantyre, Malawi.
}
\author{
\strong{Maintainer}: Lars Schöbitz \email{lschoebitz@ethz.ch} (\href{https://orcid.org/0000-0003-2196-5015}{ORCID})

Authors:
\itemize{
  \item Elizabeth Tilley \email{tilleye@ethz.ch} (\href{https://orcid.org/0000-0002-2095-9724}{ORCID})
  \item Marc Kalina \email{mkalina@ethz.ch} (\href{https://orcid.org/0000-0002-6335-3845}{ORCID})
  \item Jonathan Kwangulero \email{Kwangerlo@gmail.com} (\href{https://orcid.org/0000-0002-4050-5121}{ORCID})
  \item Hope Chilunga \email{hchilunga@poly.ac.mw} (\href{https://orcid.org/0000-0003-4216-6319}{ORCID})
}

Other contributors:
\itemize{
  \item Global Health Engineering, ETH Zurich [funder]
}

}
\keyword{internal}
