% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qechairquality.R
\docType{data}
\name{qechairquality}
\alias{qechairquality}
\title{Air quality meausurements at QECH}
\format{
A tibble with 203,806 rows and 5 variables:
\describe{
\item{date_time}{Timestamp of the air quality measurement.}
\item{location}{Label for sensor location.}
\item{indicator}{Air quality measurement indicator (pm2.5 and pm10).}
\item{value}{Measured value in µg/m3.}
}
}
\usage{
qechairquality
}
\description{
Measurements in 5-minute intervals for particulate matter (PM 2.5 & PM 10) at
eight sensor locations over two months at Queen Elizabeth Central Hospital
(QECH) in Blantyre Malawi.
}
\keyword{datasets}
